requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
bidict==0.23.1
click==8.3.1
colorama==0.4.6
fastapi==0.123.5
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pymodbus==3.11.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.0
pytz==2025.2
PyYAML==6.0.3
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
xlsxwriter==3.2.9
